Add files via upload
Write a C program that takes as input a fully parenthesized, arithmetic expression of binary
operators +, −, ∗, /, and converts the expression into a binary expression tree. Your program
should take input from the command line. The entire expression should be in a character string
without any space in it.
An input string only includes floating numbers in the format of Y.YY, that is, one digit to the left of
the decimal point and two digits to the right of the decimal point, and variables of the form of x1,
x2, ....
Your program shall allow for the leaves in the expression tree not only to store floating values but
also to store variables of the form x1, x2, x3, ..., which are initially 0.0 and can be updated
interactively by the user.

Your program should then show a menu with the following options:

1. Display
2. Preorder
3. Inorder
4. Postorder
5. Update
6. Calculate
7. Exit